Update OOPs based Java Project
This Java program is based on Object-Oriented Programming (OOP) concepts, which is a programming paradigm that is centred around objects and the data and behaviour that they contain.

The program defines two classes, Player and Team, to represent a football player and a team, respectively.

The Player class has the following instance variables:
name: a string to store the player's name.
position: a string to store the player's position.
age: an integer to store the player's age.
goals: an integer to store the number of goals the player has scored.

To return the values of these instance variables, The class also has four getter methods, 
getName()
getPosition()
getAge()
getGoals() 
	
The Team class has the following instance variables:
name: a string to store the team's name.
players: an array of Player objects to store the players in the team.


The class also has two getter methods, getName() and getPlayers(), to return the team's name and players, respectively.

In the main method of the Main class, we create two Team objects to represent Barcelona and Real Madrid and store them in an array of Team objects.

We then use a Scanner object to get input from the user. First, we display a list of teams to the user and ask them to select a team. We then retrieve the selected team from the array of teams based on the user's input and store it in a Team object.

Next, we display a list of players in the selected team to the user and ask them to select a player. We then retrieve the selected player from the array of players based on the user's input and store it in a Player object.

Finally, we display the name, position, age, and goals of the selected player using the getter methods in the Player class.